# [WIP] optimize TLV encoding-decoding