--- a/dv/dv/router.go
+++ b/dv/dv/router.go
@@ -194,8 +194,8 @@ func (dv *Router) configureFace() (err error) {
 		Module: "faces",
 		Cmd:    "update",
 		Args: &mgmt.ControlArgs{
-			Mask:  utils.IdPtr(mgmt.FaceFlagLocalFieldsEnabled),
-			Flags: utils.IdPtr(mgmt.FaceFlagLocalFieldsEnabled),
+			Mask:  enc.Some(mgmt.FaceFlagLocalFieldsEnabled),
+			Flags: enc.Some(mgmt.FaceFlagLocalFieldsEnabled),
 		},
 		Retries: -1,
 	})
@@ -246,8 +246,8 @@ func (dv *Router) register() (err error) {
 			Cmd:    "register",
 			Args: &mgmt.ControlArgs{
 				Name:   prefix,
-				Cost:   utils.IdPtr(uint64(0)),
-				Origin: utils.IdPtr(config.NlsrOrigin),
+				Cost:   enc.Some(uint64(0)),
+				Origin: enc.Some(config.NlsrOrigin),
 			},
 			Retries: -1,
 		})
@@ -278,14 +278,14 @@ func (dv *Router) register() (err error) {
 // createFaces creates faces to all neighbors.
 func (dv *Router) createFaces() {
 	for i, neighbor := range dv.config.Neighbors {
-		var mtu *uint64 = nil
+		var mtu enc.Optional[uint64]
 		if neighbor.Mtu > 0 {
-			mtu = utils.IdPtr(neighbor.Mtu)
+			mtu = enc.Some(neighbor.Mtu)
 		}
 
 		faceId, created, err := dv.nfdc.CreateFace(&mgmt.ControlArgs{
 			Uri:             utils.IdPtr(neighbor.Uri),
-			FacePersistency: utils.IdPtr(uint64(mgmt.PersistencyPermanent)),
+			FacePersistency: enc.Some(uint64(mgmt.PersistencyPermanent)),
 			Mtu:             mtu,
 		})
 		if err != nil {
@@ -304,9 +304,9 @@ func (dv *Router) createFaces() {
 			Cmd:    "register",
 			Args: &mgmt.ControlArgs{
 				Name:   dv.config.AdvertisementSyncActivePrefix(),
-				Cost:   utils.IdPtr(uint64(1)),
-				Origin: utils.IdPtr(config.NlsrOrigin),
-				FaceId: utils.IdPtr(faceId),
+				Cost:   enc.Some(uint64(1)),
+				Origin: enc.Some(config.NlsrOrigin),
+				FaceId: enc.Some(faceId),
 			},
 			Retries: 3,
 		})
@@ -322,14 +322,14 @@ func (dv *Router) destroyFaces() {
 
 		dv.engine.ExecMgmtCmd("rib", "unregister", &mgmt.ControlArgs{
 			Name:   dv.config.AdvertisementSyncActivePrefix(),
-			Origin: utils.IdPtr(config.NlsrOrigin),
-			FaceId: utils.IdPtr(neighbor.FaceId),
+			Origin: enc.Some(config.NlsrOrigin),
+			FaceId: enc.Some(neighbor.FaceId),
 		})
 
 		// only destroy faces that we created
 		if neighbor.Created {
 			dv.engine.ExecMgmtCmd("faces", "destroy", &mgmt.ControlArgs{
-				FaceId: utils.IdPtr(neighbor.FaceId),
+				FaceId: enc.Some(neighbor.FaceId),
 			})
 		}
 	}