--- a/dv/table/fib.go
+++ b/dv/table/fib.go
@@ -5,7 +5,6 @@ import (
 	"github.com/named-data/ndnd/dv/nfdc"
 	enc "github.com/named-data/ndnd/std/encoding"
 	mgmt "github.com/named-data/ndnd/std/ndn/mgmt_2022"
-	"github.com/named-data/ndnd/std/utils"
 )
 
 type FibEntry struct {
@@ -116,8 +115,8 @@ func (fib *Fib) UpdateH(nameH uint64, name enc.Name, newEntries []FibEntry) bool
 				Cmd:    "unregister",
 				Args: &mgmt.ControlArgs{
 					Name:   name,
-					FaceId: utils.IdPtr(oldEntry.FaceId),
-					Origin: utils.IdPtr(config.NlsrOrigin),
+					FaceId: enc.Some(oldEntry.FaceId),
+					Origin: enc.Some(config.NlsrOrigin),
 				},
 				Retries: 3,
 			})
@@ -138,9 +137,9 @@ func (fib *Fib) UpdateH(nameH uint64, name enc.Name, newEntries []FibEntry) bool
 			Cmd:    "register",
 			Args: &mgmt.ControlArgs{
 				Name:   name,
-				FaceId: utils.IdPtr(entry.FaceId),
-				Cost:   utils.IdPtr(entry.Cost),
-				Origin: utils.IdPtr(config.NlsrOrigin),
+				FaceId: enc.Some(entry.FaceId),
+				Cost:   enc.Some(entry.Cost),
+				Origin: enc.Some(config.NlsrOrigin),
 			},
 			Retries: 3,
 		})