--- a/dv/tlv/zz_generated.go
+++ b/dv/tlv/zz_generated.go
@@ -999,6 +999,7 @@ func (context *PrefixOpListParsingContext) Parse(reader enc.WireView, ignoreCrit
 					handled = true
 					handled_PrefixOpReset = true
 					value.PrefixOpReset = true
+					err = reader.Skip(int(l))
 				}
 			case 304:
 				if true {