--- a/std/types/optional/optional.go
+++ b/std/types/optional/optional.go
@@ -1,4 +1,4 @@
-package encoding
+package optional
 
 type Optional[T any] struct {
 	value T